USe syscall package instead of unix in test.
